src/test/java/com/github/parkerwy/CleanMojoTest.java
package com.github.parkerwy;


import org.apache.maven.plugin.testing.MojoRule;
import org.junit.Rule;
import org.junit.Test;

import java.io.File;

import static org.junit.Assert.assertNotNull;
import static org.junit.Assert.assertTrue;

public class CleanMojoTest {
    @Rule
    public MojoRule rule = new MojoRule() {
        @Override
        protected void before() throws Throwable {
        }

        @Override
        protected void after() {
        }
    };

    /**
     * @throws Exception if any
     */
    @Test
    public void shouldRunCleanMojo() throws Exception {
        File pom = new File("target/test-classes/project-to-test/");
        assertNotNull(pom);
        assertTrue(pom.exists());

        CleanMojo cleanMojo = (CleanMojo) rule.lookupConfiguredMojo(pom, "clean");
        assertNotNull(cleanMojo);
        cleanMojo.execute();
    }

}

